# historian/backend/src/main/swagger/api-swagger.yml
swagger: '2.0'
info:
  description: REST API for logisland historian
  version: v1
  title: Logisland Data Historian
  contact:
    name: hurence
    email: support@hurence.com
host: localhost:8701
basePath: /api/v1
schemes:
  - http
  - https
paths:
  /admin:
     post:
       tags:
        - tag
        - datasource
        - admin
       summary:  create new dataset with some sample data
       description:  create new dataset with some sample data
       operationId: sampleData
       produces:
          - application/json
       parameters:
         - name: flush
           in: query
           description: do we flush previous entries ?
           required: false
           type: boolean
           default: false
       responses:
         '200':
           description: Sample tag list
           schema:
             type: array
             items:
               $ref: '#/definitions/Tag'
         default:
           description: unexpected error
           schema:
             $ref: '#/definitions/Error'

  # Datasources
  /datasources:
     get:
       tags:
        - datasource
       operationId: getAllDatasources
       summary:  get all datasources
       description: retrieve all datasources
       parameters:
         - name: fq
           in: query
           description: filter query (lucene syntax like fq="labels:opc AND datasources:win32")
           required: false
           type: string
       produces:
          - application/json
       responses:
         '200':
           description: Datasource list
           schema:
             type: array
             items:
               $ref: '#/definitions/Datasource'
         default:
           description: unexpected error
           schema:
             $ref: '#/definitions/Error'
  /datasources/tags:
     get:
       tags:
        - datasource
       operationId: getAllDatasourcesTags
       summary:  get all tags from datasources
       description: retrieve all tags through this datasources
       produces:
          - application/json
       responses:
         '200':
           description: Tag list
           schema:
             type: array
             items:
               $ref: '#/definitions/Tag'
         default:
           description: unexpected error
           schema:
             $ref: '#/definitions/Error'
  /datasources/{datasourceId}/tags:
     get:
       tags:
        - tag
        - datasource
       operationId: getAllTagsFromDatasource
       summary:  get all tags from a given datasource
       description: retrieve all tags through this datasource
       produces:
          - application/json
       parameters:
         - name: datasourceId
           in: path
           description: id of the Datasource to return
           required: true
           type: string
       responses:
         '200':
           description: Tag list
           schema:
             type: array
             items:
               $ref: '#/definitions/Tag'
         default:
           description: unexpected error
           schema:
             $ref: '#/definitions/Error'
  /datasources/{datasourceId}:
     get:
       tags:
         - datasource
       summary: get Datasource
       description: get the corresponding Datasource
       operationId: getDatasource
       produces:
         - application/json
       parameters:
         - name: datasourceId
           in: path
           description: id of the Datasource to return
           required: true
           type: string
       responses:
         '200':
           description: Datasource found
           schema:
             $ref: '#/definitions/Datasource'
         '404':
            description: Datasource resource not found
         default:
           description: unexpected error
           schema:
             $ref: '#/definitions/Error'
     delete:
       tags:
         - datasource
       summary: delete Datasource
       description: remove the corresponding Datasource
       operationId: deleteDatasource
       parameters:
         - name: datasourceId
           in: path
           description: id of the Datasource to be deleted
           required: true
           type: string
       responses:
         '404':
           description: Datasource resource not found
         '400':
           description: Invalid ID supplied
         '200':
           description: Datasource successfully removed
           schema:
             $ref: '#/definitions/Datasource'
         default:
           description: unexpected error
           schema:
             $ref: '#/definitions/Error'
     put:
       tags:
        - datasource
       summary:  update Datasource
       description: update an existing Datasource
       operationId: updateDatasource
       parameters:
         - name: datasourceId
           in: path
           description: datasourceId to be updated
           required: true
           type: string
         - name: datasource
           in: body
           description: new Datasource definition
           required: true
           schema:
             $ref: '#/definitions/Datasource'
       responses:
        '404':
          description: Datasource resource not found
        '400':
          description: Invalid ID supplied
        '200':
         description: Datasource successfuly updated
         schema:
           $ref: '#/definitions/Datasource'
     post:
       tags:
        - datasource
       summary:  create new Datasource
       description:  store a new Datasource
       operationId: addDatasourceWithId
       parameters:
         - name: body
           in: body
           description: Datasource resource to add
           required: true
           schema:
             $ref: '#/definitions/Datasource'
         - name: datasourceId
           in: path
           description: datasourceId to
           required: true
           type: string
       responses:
         '400':
           description: Invalid ID supplied
         '200':
           description: Datasource successfuly created
           schema:
             $ref: '#/definitions/Datasource'
         default:
           description: unexpected error
           schema:
             $ref: '#/definitions/Error'





   # Tags API
...
  /alerts:
     get:
      tags:
        - alerts
      summary: get all alerts
      description: get the alerts
      operationId: getAlerts
      parameters:
        - name: count
          in: query
          description: max number of ites to retrieve
          required: false
          type: integer
          default: 20
        - name: severity
          in: query
          description: severity level (the higher the most severe)
          required: false
          type: integer
          default: 0
        - name: start
          in: query
          description: lower date range
          required: false
          type: string
          default: '1DAYS-AGO'
        - name: end
          in: query
          description: upper date range
          required: false
          type: string
          default: 'NOW'
      responses:
        "200":
          description: job metrics
          schema:
            type: array
            items:
              $ref: '#/definitions/Alert'
        default:
          description: unexpected error
          schema:
            $ref: '#/definitions/Error'


  # Pipeline API
...
definitions:
...
  Alert:
    type: object
    properties:
      record_type:
        type: string
        default: 'alert'
      timestamp:
        type: integer
        format: int64
      message:
        type: string
      severity:
        type: integer
      tag_id:
        $ref: '#/definitions/Tag'
  Datasource:
    type: object
    required:
      - id
      - host
    properties:
      record_type:
        type: string
        default: 'datasource'
      id:
        type: string
      description:
        type: string
      clsid:
        type: string
      datasource_type:
        type: string
      prog_id:
        type: string
      host:
        type: string
      domain:
        type: string
      user:
        type: string
      password:
        type: string
...
  Tag:
    type: object
    description: a Tag is an identifier to an OPC value
    required:
      - item_name
      - datasource_id
      - data_type
    properties:
      record_type:
        type: string
        default: 'tag'
      id:
        type: string
        default: 'mySweetUniqueId'
      datasource_id:
        type: string
      domain:
        type: string
      server:
        type: string
      group:
        type: string
      tag_name:
        type: string
      labels:
        description: can be used to group tags around a theme
        type: array
        items:
          type: string
      data_type:
        type: string
        default: float
        enum:
        - int
        - long
        - float
        - double
        - string
        - array
        - null
        - bytes
        - boolean
      description:
        type: string
      text:
        description: ctach all field
        type: array
        items:
          type: string
      creation_date:
        description: timestamp
        type: integer
        format: int64
      last_modification_date:
        description: timestamp
        type: integer
        format: int64
      last_polling_date:
        description: timestamp
        type: integer
        format: int64
      update_rate:
        description: polling delay in ms
        default: 1000
        type: integer
      min_numeric_value:
        type: number
        format: dumble
      max_numeric_value:
        type: number
        format: double
      last_numeric_value:
        type: number
        format: double
      last_quality:
        type: integer
...
  Error:
    required:
      - code
      - message
    properties:
      code:
        type: integer
        format: int32
      message:
        type: string